:wrench: refactor : 마이페이지 light house를 이용한 최적화 작업 (#218)
