+++ src/app/(after-login)/(background-gray)/mypage/page.tsx
@@ -1,50 +1,18 @@
-'use client';
-
-import { usePostImage } from '@/apis/image/queries';
-import LogoutButton from '@/components/auth/LogoutButton';
 import DailyEmotion from '@/components/mypage/DailyEmotion';
 import EmotionCalendar from '@/components/mypage/EmotionCalendar';
-import EmotionChart from '@/components/mypage/EmotionChart';
+import EmotionChartSection from '@/components/mypage/EmotionChartSection';
 import MyEpigramComments from '@/components/mypage/MyEpigramComments';
-import Avatar from '@/components/ui/avatars';
-import { MypageContext } from '@/context/MypageProvider';
-import { useContext, useRef } from 'react';
+import UserProfileLayout from '@/components/mypage/UserProfileLayout';
 
 export default function MyPage() {
-  const { user } = useContext(MypageContext);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const { mutate: upLoadImage } = usePostImage();
-
-  const handleAvatarClick = () => {
-    fileInputRef.current?.click();
-  };
-
-  const handleChangeImage = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-
-    const formData = new FormData();
-    formData.append('image', file);
-
-    upLoadImage(formData);
-  };
-
   return (
     <div className='bg-background-100 pt-[116px] lg:pt-[208px]'>
       <div className='shadow-custom relative flex flex-col items-center justify-center rounded-3xl bg-blue-100 px-6 pt-[184px] pb-10 shadow-black md:pb-[63px] lg:pt-[276px] lg:pb-22'>
-        <div className='absolute top-0 left-1/2 flex -translate-x-1/2 -translate-y-[40px] flex-col items-center justify-center gap-4 lg:-translate-y-[60px] lg:gap-6'>
-          <Avatar src={user.image} alt='유저이미지' className='size-20 cursor-pointer border-2 border-blue-300 lg:size-[120px]' onClick={handleAvatarClick}></Avatar>
-          <input type='file' className='hidden' onChange={handleChangeImage} accept='image/*' ref={fileInputRef} />
-          <p className='text-black-950 text-lg font-medium lg:text-2xl'>{user.nickname}</p>
-          <LogoutButton className='h-9 w-[77px] px-3.5 py-1.5 text-[14px] font-normal whitespace-nowrap lg:h-12 lg:w-[100px] lg:px-4 lg:py-2 lg:text-xl lg:font-medium'></LogoutButton>
-        </div>
+        <UserProfileLayout></UserProfileLayout>
         <div className='flex flex-col gap-14 md:gap-15 lg:gap-41'>
           <DailyEmotion></DailyEmotion>
           <EmotionCalendar></EmotionCalendar>
-          <div className='flex flex-col gap-4 lg:gap-12'>
-            <h2 className='text-lg font-semibold lg:text-2xl'>감정 차트</h2>
-            <EmotionChart></EmotionChart>
-          </div>
+          <EmotionChartSection></EmotionChartSection>
         </div>
       </div>
       <div className='mx-auto w-[360px] pt-14 pb-[78px] md:w-[384px] md:pb-[102px] lg:w-[640px] lg:pt-24 lg:pb-[156px]'>
